Trait trillium_aws_lambda::LambdaConnExt

pub trait LambdaConnExt {
    fn lambda_context(&self) -> &Context;
}

Provides access to the aws lambda context for trillium::Conn.

See [lamedh_runtime::Context] for more details on the data available on this struct.

Required methods

fn lambda_context(&self) -> &Context

returns the [lamedh_runtime::Context] for this conn